requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.13.2
pydantic==2.9.2
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4